Organize components
Organize components that involve languages for comparison feature
