+++ src/components/home/comparison/Languages/BarGraphLangCanada.jsx
