+++ src/components/home/comparison/Languages/BarGraphLangUSA.jsx
